py-client/requirements.txt
Beaker==1.11.0
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
cryptography==37.0.4
defusedxml==0.7.1
future==0.18.2
humanfriendly==10.0
idna==3.3
Mako==1.2.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oic==1.4.0
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
pycryptodomex==3.15.0
pyjwkest==1.4.2
PyJWT==2.4.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
